requirements.txt
streamlit
pandas
altair
mysqlclient
SQLAlchemy
mysql-connector-python
